Fix newlines_after_classes when class ends with indent
LineApi.maintainClassContext didn't work properly when class definition ended
with an indent, as only 1 occurrence of INDENT and OUTDENT was countend per line.
This meant that if one line OUTDENT-end multiple INDENTs the
maintainClassContext routine failed to detect the class ended.

Fixed by properly counting the amount of occurrences of INDENT and OUTDENTs.